Refactor (#71)
* Replace code to use uuid string constructors

* uuid str replaced with byte for compatibility

* redo byte reading in SceneItemModel.cpp

* more profiling

* Scene opening is now post update

* step gc

* If scene was in playmode open new scene in playmode

* Minor performance improvements, hopefully

* tiny fixes after merging

* Replace uuid_v4 with uint64_t uuid

* Delete editor/ui/scene/SceneItemModel.cpp

* Delete editor/src/scene/SceneMeta.cpp

* change to underlying type, reduce amount of magic numbers

* Pretty up lua type registration

* mp flag and no more 120 symbols in a line

* untrack cmake presets in git

* warnings as errors

* Logger cleanup. Add clang-tidy file

* load textures and models asynchronously

* fix timeline crash

* fix reparanting crashes

* Bugprone clang-tidy errors fixes

---------

+++ engine/include/tools/FreeListVector.h
@@ -118,7 +118,7 @@ template <typename T> void FreeListVector<T>::erase(size_t index)
     if (index < m_data.size())
     {
         m_data[index].~T();
-        m_freeList.push_back(index);
+        m_freeList.push_back(static_cast<unsigned int>(index));
     }
 }
 
@@ -137,7 +137,7 @@ template <typename T> bool FreeListVector<T>::is_occupied(size_t index)
 {
     if (index >= m_data.size()) return false;
 
-    return eastl::find(m_freeList.begin(), m_freeList.end(), index) == m_freeList.end();
+    return eastl::find(m_freeList.begin(), m_freeList.end(), static_cast<unsigned int>(index)) == m_freeList.end();
 }
